feat: refactor Split PDF success page with loading state and new content component
