+++ src/app/split-pdf/success/components/NoResult.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import {Header} from "@/components/Header";
 import {Card, CardContent, CardHeader, CardTitle} from "@/components/ui/card";
 import {Button} from "@/components/ui/button";
